# Improve performance of form-decode