Improves performance by avoiding regular expressions and allocation of
immutable collections where possible