@weavejester 2021-08-30
What did you benchmark this change to be in comparison to the existing code? How does it compare to splitting on a string? What happens if a tokenizer is used onto to split on `&` and not `=`?