@bsless 2021-08-31
- Benchmarks: I used the examples in tests as test cases, and some contrived cases such as `a=1&b=2&c=3&d=4&e=5`. I don't mind rerunning the benchmarks and sharing the results here. If there are some pathological cases which you think might perform badly I would like to bench them as well. I also used clj-async-profiler to generate flame graphs.
- Tokenizing vs. splitting on a string: this has two levels. One is that splitting uses regular expressions which dominate the flame graphs while tokenization uses string comparison. The other is that `str/split` also allocates a vector which is later traversed by reduce. Iterating over an array with indexed access would have been faster. Thus, the implementation avoids regular expressions, dispatching via protocols, destructuring and some allocations.
- Not sure I understand the last question, can you please clarify?