@weavejester 2021-08-31
I had it in my head that `String.split` had an overload for splitting on a `char` (and therefore avoiding the overhead of a regex), but I was thinking of `String.replace`. Sorry for the confusion.

I'd appreciate the benchmark results, as it helps document the decisions made in the project. We can trace a commit back to a PR, and from the PR we can see how much impact a particular performance change made.