@bsless 2021-08-31
Apologies for the delay. Here are the changes, presented nicely and reproducible:
```clojure
(ns user
  (:require
   [ring.util.codec :refer [form-decode]]
   [criterium.core :as cc]))


(def forms
  ["foo"
   "a=b"
   "a=b&c=d"
   "foo+bar"
   "a=b+c"
   "a=b%2Fc"
   "a=b&c"
   "a=&b=c"
   "a&b=c"
   "="
   "a="
   "=b"
   "a=1&b=2&c=3&d=4"
   "a=1&b=2&c=3&d=4&x=1&y=2&z=3&u=4"
   "aaaaaaaaaaaaa=111111111111111&bbbbbbbbbbbbbbb=2222222222222"])

(doseq [form forms]
  (let [res (first (:mean (cc/quick-benchmark (form-decode form) {})))
        [scale unit] (cc/scale-time res)
        mean (* scale res)]
    (println form mean unit)))


;;; foo                             16.165265103567577 ns -> 18.071835515056534 ns
;;; a=b                             281.7666082195549  ns -> 85.34618203356219  ns
;;; a=b&c=d                         520.3155533101125  ns -> 192.88535391228297 ns
;;; foo+bar                         28.89783207896534  ns -> 33.96275410375853  ns
;;; a=b+c                           288.6556301231446  ns -> 100.12228382735458 ns
;;; a=b%2Fc                         323.94066534536466 ns -> 129.9906610438482  ns
;;; a=b&c                           470.7898489852522  ns -> 209.38083310542123 ns
;;; a=&b=c                          561.700251137112   ns -> 204.26021604451992 ns
;;; a&b=c                           481.1326203353287  ns -> 198.07185181140767 ns
;;; =                               252.35371110995692 ns -> 72.2746499966358   ns
;;; a=                              263.9456664837952  ns -> 84.74253809189693  ns
;;; =b                              262.19896201858967 ns -> 85.57113803290105  ns
;;; a=1&b=2&c=3&d=4                 1.0210454463867895 µs -> 509.19558470486334 ns
;;; a=1&b=2&c=3&d=4&x=1&y=2&z=3&u=4 2.0973807618522025 µs -> 1.166763297769531  µs

;;; aaaaaaaaaaaaa=111111111111111&bbbbbbbbbbbbbbb=2222222222222
;;; 840.1896265647041  ns -> 471.0025293913937  ns
```