@weavejester 2021-09-02
Let's see if we can't factor out the tokenizer a little. What if we created a reducible:

```clojure
(defn tokenized [s delim]
  (reify clojure.lang.IReduceInit
    (reduce [_ f init]
      (let [tokenizer (StringTokenizer. s delim)]
        (loop [result init]
          (if (.hasMoreTokens tokenizer)
            (recur (f result (.nextToken tokenizer)))
            result))))))
```

Then we can just replace `(str/split encoded #"&")` with `(tokenized encoded "&")` in the original code.