@bsless 2021-09-02
This would only address the outer tokenizer loop, though.
Pros:
- Probably more readable
- Less code changes at the code site
Cons:
- slightly more allocations
- dispatching through `reduce`
- If we want the implementation to be correct we should add a call to `reduced?` every iteration

I don't find tokenizers and interop too noisy today. If the goal is enhancing readability, there are a few alternatives:

1. Reduce externally, keep internal loop the same: (see note in the end)

```clojure
(defn form-decode
  "Decode the supplied www-form-urlencoded string using the specified encoding,
  or UTF-8 by default. If the encoded value is a string, a string is returned.
  If the encoded value is a map of parameters, a map is returned."
  ([encoded]
   (form-decode encoded "UTF-8"))
  ([^String encoded encoding]
   (if-not (.contains encoded "=")
     (form-decode-str encoded encoding)
     (reduce
      (fn [m param]
        (let [param-tokenizer (StringTokenizer. param "=")
              k (if (.startsWith param "=")
                  ""
                  (if (.hasMoreTokens param-tokenizer)
                    (.nextToken param-tokenizer)
                    ""))
              v (if (.hasMoreTokens param-tokenizer) (.nextToken param-tokenizer) "")
              k (form-decode-str k encoding)
              v (form-decode-str v encoding)]
          (if (and k v)
            (assoc-conj m k v)
            m)))
      {}
      (tokenized encoded #"&")))))
```

2. Try to improve readability around most optimal code:
This solution emits exactly the same bytecode as the original

```clojure
(definline tokenizer
  [^String s ^String token]
  `(StringTokenizer. ~s ~token))

(definline has-more?
  [^StringTokenizer st]
  `(.hasMoreTokens ~st))

(defn next-token
  {:inline
   (fn
     ([st] `(.nextToken ~st))
     ([st nf] `(if (has-more? ~st) (.nextToken ~st) ~nf)))
   :inline-arities #{1 2}}
  ([^StringTokenizer st]
   (.nextToken st))
  ([^StringTokenizer st nf]
   (if (has-more? st) (.nextToken st) nf)))

(defn form-decode
  "Decode the supplied www-form-urlencoded string using the specified encoding,
  or UTF-8 by default. If the encoded value is a string, a string is returned.
  If the encoded value is a map of parameters, a map is returned."
  ([encoded]
   (form-decode encoded "UTF-8"))
  ([^String encoded encoding]
   (if-not (.contains encoded "=")
     (form-decode-str encoded encoding)
     (let [tokenizer (tokenizer encoded "&")]
       (loop [m {}]
         (if (has-more? tokenizer)
           (let [^String param (next-token tokenizer)
                 ^StringTokenizer param-tokenizer (tokenizer param "=")
                 k (if (.startsWith param "=")
                     ""
                     (next-token param-tokenizer ""))
                 k (form-decode-str k encoding)
                 v (form-decode-str (next-token param-tokenizer "") encoding)]
             (recur (if (and k v) (assoc-conj m k v) m)))
           m))))))
```

3. Reduce externally, keep old function body (defeats the purpose in my opinion, as the highest cost we pay here is over regular expressions)

NOTE:

Performance differences for reified reducible tokenizer: (solution 1)

```
;;; foo                             16.165265103567577 ns -> 18.071835515056534 ns -> 17.320719336705146 ns
;;; a=b                             281.7666082195549  ns -> 85.34618203356219  ns -> 100.27655876859055 ns
;;; a=b&c=d                         520.3155533101125  ns -> 192.88535391228297 ns -> 231.03698842087675 ns
;;; foo+bar                         28.89783207896534  ns -> 33.96275410375853  ns -> 29.902639183635433 ns
;;; a=b+c                           288.6556301231446  ns -> 100.12228382735458 ns -> 113.89209010375781 ns
;;; a=b%2Fc                         323.94066534536466 ns -> 129.9906610438482  ns -> 136.2942337242038  ns
;;; a=b&c                           470.7898489852522  ns -> 209.38083310542123 ns -> 210.61152438110452 ns
;;; a=&b=c                          561.700251137112   ns -> 204.26021604451992 ns -> 224.44069615994098 ns
;;; a&b=c                           481.1326203353287  ns -> 198.07185181140767 ns -> 220.36356595037435 ns
;;; =                               252.35371110995692 ns -> 72.2746499966358   ns -> 83.88891707876553  ns
;;; a=                              263.9456664837952  ns -> 84.74253809189693  ns -> 98.03556783118877  ns
;;; =b                              262.19896201858967 ns -> 85.57113803290105  ns -> 98.35710802210014  ns
;;; a=1&b=2&c=3&d=4                 1.0210454463867895 µs -> 509.19558470486334 ns -> 529.099610694873   ns
;;; a=1&b=2&c=3&d=4&x=1&y=2&z=3&u=4 2.0973807618522025 µs -> 1.166763297769531  µs -> 1.1895777987654739 µs

;;; aaaaaaaaaaaaa=111111111111111&bbbbbbbbbbbbbbb=2222222222222
;;; 840.1896265647041  ns -> 471.0025293913937  ns -> 481.37886251787927 ns
```

We can consistently see a slight overhead, so in terms of performance there's no great impact.
I don't believe the alternative implementation improves readability, though.

What do you think?